requirements.txt
torch
torchvision
coremltools>=7.0
Pillow
numpy
